Create interactive eyes on black canvas
This commit adds p5.js and p5-react-wrapper packages to allow for
image and shape manipulation functionality.

This is then used to create a black canvas with two eyes that follow
the mouse. On mouse click, an arc is added to make the eyes look angry.

Most unused boilerplate code has been removed.