Add touchscreen toast prompting touch for audio
Due to browser policies, audio can only be played once the user
has interacted in the form of a click or touch event. On touchscreen
devices, only a quick tap satisfies this requirement, but holding or
dragging does not.

For this reason, a brief "toast" has been added to instruct the user
to tap once to enable audio.